% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{dataDWDPerDay.old}
\alias{dataDWDPerDay.old}
\title{dataDWDPerDay.old}
\usage{
dataDWDPerDay.old(date = Sys.Date() - 1, hour = "23")
}
\arguments{
\item{date}{as Date formate YY-MM-DD}

\item{hour}{hour of the day}
}
\value{
a radolan grid
}
\description{
This function lists all available sensors for a specific device on open sensor web
}
\examples{
radolanr::dataDWDPerDay(date = Sys.Date()-1, hour = "23")
}
