% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{dataDWDPerDay}
\alias{dataDWDPerDay}
\title{dataDWDPerDay}
\usage{
dataDWDPerDay(mode = "specific", date = Sys.Date() - 1, hour = "23")
}
\arguments{
\item{mode}{latest or specific if only the last dataset provided shall be used, variables date and hour will be ignored an the last dataset provided will be returned, the https server might also provide this but can not be accessed in the same way with rdwd}

\item{date}{as Date formate YY-MM-DD}

\item{hour}{hour of the day}
}
\value{
a radolan grid
}
\description{
This function lists all available sensors for a specific device on open sensor web
A file like "lastest" is only available at https but data is only available for 2 days at "https", so I will stay with ftp but try to find the last dataset in mode = "latest"
I the code is run on a system with another timezone the the one of the user, this might cause problems
This function and rdwd is not using #https://opendata.dwd.de/weather/radar/radolan/sf/ but gridbase = "ftp://opendata.dwd.de/climate_environment/CDC/grids_germany"
For Radolan finaly this function then is using: radbase = "ftp://opendata.dwd.de/climate_environment/CDC/grids_germany/daily/radolan/recent/bin/"
}
\examples{
radolanr::dataDWDPerDay(date = Sys.Date()-1, hour = "23")
}
