% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{dataDWDPerDayInterval}
\alias{dataDWDPerDayInterval}
\title{dataDWDPerDayInterval}
\usage{
dataDWDPerDayInterval(
  date.start = as.Date("2025-01-01"),
  date.end = as.Date("2025-01-02"),
  hour = "23",
  ...
)
}
\arguments{
\item{date.end}{as.Date format e.g. as.Date("2025-01-01")}

\item{hour}{hour of the day}

\item{date}{as Date formate YY-MM-DD}

\item{data.start}{as.Date format}
}
\value{
a radolan grid
}
\description{
Download data of Radolan Data for a specific time interval
all arguments from dataDWDPerDay() may be used to e.g. specify the reference hour for the measurement, default is 23:00, but data is from the last 24h, so 23:00 largly represents the precipitation of the given day.
}
\examples{
radolanr::dataDWDPerDay(date = Sys.Date()-1, hour = "23")
}
