% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{extractTimeSeriesFromStack}
\alias{extractTimeSeriesFromStack}
\title{extractTimeSeriesFromStack}
\usage{
extractTimeSeriesFromStack(raster_stack, centroids)
}
\arguments{
\item{raster_stack}{given radolan raster stack, exportet from dataDWDPerDayInterval()}

\item{centroids}{in the format of SpatialPoints for which data shall be exported}
}
\value{
a list of the resulting timeseries as data.frame along with given centroids and raster stack
}
\description{
extract a timeseries object from a radolan-stack
}
