% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{floor_time_to_10min}
\alias{floor_time_to_10min}
\title{floor_time_to_10min}
\usage{
floor_time_to_10min(mytime)
}
\arguments{
\item{raster_stack}{given radolan raster stack, exportet from dataDWDPerDayInterval()}

\item{centroids}{in the format of SpatialPoints for which data shall be exported}
}
\value{
a list of the resulting timeseries as data.frame along with given centroids and raster stack
}
\description{
round_time_to_10min
}
\examples{
my.centroids = terra::vect("data_raw/centroid_amms_weatherstation_coswig_4326.shp")
my.timeseries <- extractTimeSeriesFromStack(raster_stack = radolan_stack, centroids = my.centroids)
}
