% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radolan2polygon.R
\name{radolan2polygon}
\alias{radolan2polygon}
\title{radolan2polygon.R}
\usage{
radolan2polygon(
  locations_polygon_path,
  radolan_raster,
  use_metadata = TRUE,
  silent = TRUE,
  saveCentroidsVector = FALSE,
  centroidsVectorPath = ""
)
}
\arguments{
\item{locations_polygon_path}{a polygon layer (.shp)}

\item{radolan_raster}{the radolan raster created with radolanr::dataDWDPerDay(mode = "latest", addMetaData = TRUE)}

\item{use_metadata}{use metadata stored in a terra::metag() in the radolan_raster only set to TRUE if dataDWDPerDay(mode = "latest", addMetaData = TRUE) was used in prior processing}

\item{silent}{show warnings, plots,..}

\item{saveCentroidsVector}{export the created centroids to a shapefile (boolean)}

\item{centroidsVectorPath}{path where the centroids will be exported to (.shp)}
}
\value{
a polygon layer with locations_polygon_path filled with rain_mm values from radolan data
}
\description{
This function
Outlook: I will in future maybe use terra alongside with geos package. Use terra if working with raster & vector data (most modern for raster). Use sf if doing vector operations (widely used, but heavier dependencies). Use geos for fast geometry computations (complements terra well).
}
\examples{
radolan2polygon(locations_polygon_path = "data_raw/transsect_weinbau_seusslitz_pillnitz_polygons_4326.shp", radolan_raster = radp)
}
