% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadolanTimeseries.R
\name{dataDWDPerDayInterval}
\alias{dataDWDPerDayInterval}
\title{dataDWDPerDayInterval}
\usage{
dataDWDPerDayInterval(
  date_start = as.Date("2025-01-01"),
  date_end = as.Date("2025-01-02"),
  hour = "23",
  ...
)
}
\arguments{
\item{date_start}{as.Date format}

\item{date_end}{as.Date format e.g. as.Date("2025-01-01")}

\item{hour}{description}

\item{...}{some other arguments}
}
\value{
a radolan grid
}
\description{
Download data of Radolan Data for a specific time interval
all arguments from dataDWDPerDay() may be used to e.g. specify the reference hour for the measurement, default is 23:00, but data is from the last 24h, so 23:00 largly represents the precipitation of the given day.
}
\examples{
radolanr::dataDWDPerDay(date = Sys.Date()-1, hour = "23")
}
